Evaluate `cfg(feature = "tracing")` in correct scope (#5)
The content of TokenStream a proc_macro returns is evaluated in the
context of its caller. I.e. any `#[cfg(…)]` will look if the crate that
uses the proc_macro fulfills the guard, and not whether the proc_macro
lib fulfills the guard.

This PR makes sure that the `#[cfg(…)]` is evaluated in the context of
the proc_macro.